some naming and cleanup

+++ components/PostItem/PostItem.tsx
@@ -1,5 +1,5 @@
 import { JSX } from 'preact'
-import { Post } from '../utils/posts.ts'
+import { Post } from '../../utils/posts.ts'
 
 export default function PostItem(props: { post: Post }) {
   const { post } = props
@@ -13,7 +13,7 @@ export default function PostItem(props: { post: Post }) {
         <time>
           {new Date(post.publishedAt).toLocaleDateString("en-us", {
             year: "numeric",
-            month: "long",
+            month: "short",
             day: "numeric",
           })}
         </time>
